# fern/definition/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Airweave",
    "version": "0.1.0",
    "description": "\n\n## API Groups\nThis API spec only includes the following API groups:\n- **Sources**: API endpoints for discovering available data source connectors and their configuration requirements\n- **Collections**: API endpoints for managing collections - logical groups of data sources that provide unified search capabilities\n- **Source Connections**: API endpoints for managing live connections to data sources. Source connections are the actual configured instances that Airweave uses to sync data from your apps and databases, transforming it into searchable, structured information within collections\n"
  },
  "paths": {
    "/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List",
        "description": "List all available data source connectors.\n\n<br/><br/>\nReturns the complete catalog of source types that Airweave can connect to.",
        "operationId": "list_sources_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Source"
                  },
                  "title": "Response List Sources Get"
                },
                "examples": {
                  "available_sources": {
                    "summary": "List of all available data source connectors",
                    "value": [
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440000",
                        "name": "GitHub",
                        "description": "Connect to GitHub repositories for code, issues, pull requests, and documentation",
                        "auth_type": "config_class",
                        "auth_config_class": "GitHubAuthConfig",
                        "config_class": "GitHubConfig",
                        "short_name": "github",
                        "class_name": "GitHubSource",
                        "output_entity_definition_ids": [
                          "def12345-6789-abcd-ef01-234567890abc",
                          "def67890-abcd-ef01-2345-67890abcdef1"
                        ],
                        "labels": [
                          "code"
                        ],
                        "created_at": "2024-01-01T00:00:00Z",
                        "modified_at": "2024-01-01T00:00:00Z",
                        "auth_fields": {
                          "fields": [
                            {
                              "name": "personal_access_token",
                              "title": "Personal Access Token",
                              "description": "GitHub Personal Access Token with repository read permissions",
                              "type": "string",
                              "secret": true
                            },
                            {
                              "name": "repo_name",
                              "title": "Repository Name",
                              "description": "Full repository name in format 'owner/repo'",
                              "type": "string"
                            }
                          ]
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/sources/{short_name}": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Get",
        "description": "Get detailed information about a specific data source connector.",
        "operationId": "get_sources__short_name__get",
        "parameters": [
          {
            "name": "short_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Technical identifier of the source type (e.g., 'github', 'stripe', 'slack')",
              "title": "Short Name"
            },
            "description": "Technical identifier of the source type (e.g., 'github', 'stripe', 'slack')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Source"
                },
                "examples": {
                  "github_source": {
                    "summary": "Source details with authentication and configuration schemas",
                    "value": {
                      "id": "550e8400-e29b-41d4-a716-446655440000",
                      "name": "GitHub",
                      "description": "Connect to GitHub repositories for code, issues, pull requests, and documentation",
                      "auth_type": "config_class",
                      "auth_config_class": "GitHubAuthConfig",
                      "config_class": "GitHubConfig",
                      "short_name": "github",
                      "class_name": "GitHubSource",
                      "output_entity_definition_ids": [
                        "def12345-6789-abcd-ef01-234567890abc",
                        "def67890-abcd-ef01-2345-67890abcdef1"
                      ],
                      "labels": [
                        "code"
                      ],
                      "created_at": "2024-01-01T00:00:00Z",
                      "modified_at": "2024-01-01T00:00:00Z",
                      "auth_fields": {
                        "fields": [
                          {
                            "name": "personal_access_token",
                            "title": "Personal Access Token",
                            "description": "GitHub Personal Access Token with repository read permissions",
                            "type": "string",
                            "secret": true
                          },
                          {
                            "name": "repo_name",
                            "title": "Repository Name",
                            "description": "Full repository name in format 'owner/repo'",
                            "type": "string"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/source-connections": {
      "post": {
        "tags": [
          "source-connections"
        ],
        "summary": "Create",
        "description": "Create a new source connection.\n\nThe authentication configuration determines the flow:\n- DirectAuthentication: Immediate creation with provided credentials\n- OAuthBrowserAuthentication: Returns shell with authentication URL\n- OAuthTokenAuthentication: Immediate creation with provided token\n- AuthProviderAuthentication: Using external auth provider\n\nBYOC (Bring Your Own Client) is detected when client_id and client_secret\nare provided in OAuthBrowserAuthentication.\n\nsync_immediately defaults:\n- True for: direct, oauth_token, auth_provider\n- False for: oauth_browser, oauth_byoc (these sync after authentication)",
        "operationId": "create_source_connections_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceConnectionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceConnection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "source-connections"
        ],
        "summary": "List",
        "description": "List source connections with minimal fields for performance.",
        "operationId": "list_source_connections_get",
        "parameters": [
          {
            "name": "collection",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by collection readable ID",
              "title": "Collection"
            },
            "description": "Filter by collection readable ID"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceConnectionListItem"
                  },
                  "title": "Response List Source Connections Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/source-connections/{source_connection_id}": {
      "get": {
        "tags": [
          "source-connections"
        ],
        "summary": "Get",
        "description": "Get a source connection with optional depth expansion.",
        "operationId": "get_source_connections__source_connection_id__get",
        "parameters": [
          {
            "name": "source_connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceConnection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "source-connections"
        ],
        "summary": "Update",
        "description": "Update a source connection.\n\nUpdateable fields:\n- name, description\n- config_fields\n- cron_schedule\n- auth_fields (direct auth only)",
        "operationId": "update_source_connections__source_connection_id__patch",
        "parameters": [
          {
            "name": "source_connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Connection Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceConnectionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceConnection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "source-connections"
        ],
        "summary": "Delete",
        "description": "Delete a source connection and all related data.",
        "operationId": "delete_source_connections__source_connection_id__delete",
        "parameters": [
          {
            "name": "source_connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceConnection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/source-connections/{source_connection_id}/run": {
      "post": {
        "tags": [
          "source-connections"
        ],
        "summary": "Run",
        "description": "Trigger a sync run for a source connection.\n\nRuns are always executed through Temporal workflow engine.\n\nArgs:\n    db: Database session\n    source_connection_id: ID of the source connection to run\n    ctx: API context with organization and user information\n    guard_rail: Guard rail service for usage limits\n    force_full_sync: If True, forces a full sync with orphaned entity cleanup\n                    for continuous syncs. Raises 400 error if used on\n                    non-continuous syncs (which are always full syncs).",
        "operationId": "run_source_connections__source_connection_id__run_post",
        "parameters": [
          {
            "name": "source_connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Connection Id"
            }
          },
          {
            "name": "force_full_sync",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Force a full sync ignoring cursor data instead of waiting for the daily cleanup schedule. Only allowed for continuous syncs.",
              "default": false,
              "title": "Force Full Sync"
            },
            "description": "Force a full sync ignoring cursor data instead of waiting for the daily cleanup schedule. Only allowed for continuous syncs."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceConnectionJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/source-connections/{source_connection_id}/jobs": {
      "get": {
        "tags": [
          "source-connections"
        ],
        "summary": "Get Source Connection Jobs",
        "description": "Get sync jobs for a source connection.",
        "operationId": "get_source_connection_jobs_source_connections__source_connection_id__jobs_get",
        "parameters": [
          {
            "name": "source_connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Connection Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceConnectionJob"
                  },
                  "title": "Response Get Source Connection Jobs Source Connections  Source Connection Id  Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/source-connections/{source_connection_id}/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "source-connections"
        ],
        "summary": "Cancel Job",
        "description": "Cancel a running sync job for a source connection.\n\nThis endpoint requests cancellation and marks the job as CANCELLING.\nThe workflow updates the final status to CANCELLED when it processes\nthe cancellation request.",
        "operationId": "cancel_job_source_connections__source_connection_id__jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "source_connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Connection Id"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceConnectionJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authentication"
      }
    }
  },
...
  "servers": [
    {
      "url": "https://api.airweave.ai",
      "description": "Production",
      "x-fern-server-name": "Production"
    },
    {
      "url": "http://localhost:8001",
      "description": "Local",
      "x-fern-server-name": "Local"
    }
  ]
}